+++ src/main/resources/templates/db_scripts/uniqueRequestItemIdAndPosition.sql
@@ -1,0 +1,2 @@
+CREATE UNIQUE INDEX request_itemid_position_idx_unique
+  ON ${myuniversity}_${mymodule}.request((jsonb->'itemId'), (jsonb->'position'));